requirements.txt
prefect>=2.0
opentelemetry-instrumentation==0.33b0
opentelemetry-instrumentation-httpx==0.33b0
opentelemetry-instrumentation-fastapi==0.33b0
opentelemetry-instrumentation-asyncpg==0.33b0
opentelemetry-instrumentation-sqlalchemy==0.33b0
# opentelemetry-instrumentation-sqlite3==0.33b0
opentelemetry-sdk==1.12.0
opentelemetry-exporter-otlp-proto-grpc==1.12.0
opentelemetry-exporter-otlp-proto-http==1.12.0
opentelemetry-api==1.12.0